chore: New `cleanup` file to store our cleanup functions
- Moved artwork cleanup function to this file.
