+++ mobile/src/modules/scanning/helpers/artwork.ts
@@ -4,18 +4,17 @@ import {
 } from "@missingcore/react-native-metadata-retriever";
 import { createId } from "@paralleldrive/cuid2";
 import { eq, inArray, isNotNull, or } from "drizzle-orm";
-import { Directory } from "expo-file-system";
 
 import { db } from "~/db";
-import { albums, artists, playlists, tracks } from "~/db/schema";
+import { albums, tracks } from "~/db/schema";
 
 import { getAlbums, updateAlbum } from "~/api/album";
 import { getTracks, updateTrack } from "~/api/track";
 import { onboardingStore } from "../services/Onboarding";
 
-import { ImageDirectory, deleteImage } from "~/lib/file-system";
+import { ImageDirectory } from "~/lib/file-system";
 import { Stopwatch } from "~/utils/debug";
-import { BATCH_PRESETS, batch } from "~/utils/promise";
+import { BATCH_PRESETS } from "~/utils/promise";
 
 type PartialTrack = {
   id: string;
@@ -24,12 +23,11 @@ type PartialTrack = {
   uri: string;
 };
 
-//#region Saving Function
 /** Save artwork for albums & tracks. */
 export async function findAndSaveArtwork() {
   const stopwatch = new Stopwatch();
 
-  // Ensure we don't unnecessarily seach for artwork.
+  // Ensure we don't unnecessarily search for artwork.
   const albumsWithCovers = await getAlbums({
     columns: ["id"],
     withTracks: false,
@@ -110,27 +108,7 @@ export async function findAndSaveArtwork() {
   );
 }
 
-/** Iterate over a list of tracks, finding and saving its artwork. */
-async function saveSinglesArtwork(
-  singles: PartialTrack[],
-  onSave: (info: { artworkUri: string; trackId: string }) => Promise<void>,
-  options?: { endEarly?: boolean; onEndIteration?: VoidFunction },
-) {
-  for (const { id: trackId, uri } of singles) {
-    // Indicate we attempted to find artwork for a track. Do this before
-    // physically attempting to save the artwork in case an OOM error occurs,
-    // in which the app will essentially become "bricked".
-    await updateTrack(trackId, { fetchedArt: true });
-    const { uri: artworkUri } = await getArtworkUri(uri);
-    if (artworkUri) {
-      await onSave({ artworkUri, trackId });
-      onboardingStore.setState((prev) => ({ found: prev.found + 1 }));
-      if (options?.endEarly) return;
-    }
-    if (options?.onEndIteration) options.onEndIteration();
-  }
-}
-
+//#region Helpers
 /**
  * Create a uri associated with the artwork on the track.
  *
@@ -151,45 +129,25 @@ export async function getArtworkUri(uri: string) {
 }
 //#endregion
 
-//#region Cleanup Function
-export async function cleanupImages() {
-  // Get all the uris of images saved in the database.
-  const usedUris = (
-    await Promise.all([
-      ...[artists, playlists].map((schema) =>
-        db
-          .select({ artwork: schema.artwork })
-          .from(schema)
-          .where(isNotNull(schema.artwork)),
-      ),
-      ...[albums, tracks].flatMap((schema) => [
-        db
-          .select({ artwork: schema.altArtwork })
-          .from(schema)
-          .where(isNotNull(schema.altArtwork)),
-        db
-          .select({ artwork: schema.embeddedArtwork })
-          .from(schema)
-          .where(isNotNull(schema.embeddedArtwork)),
-      ]),
-    ])
-  )
-    .flat()
-    .map(({ artwork }) => artwork!);
-
-  // Get & delete all unused images.
-  let deletedCount = 0;
-  await batch({
-    data: new Directory(ImageDirectory)
-      .list()
-      // There shouldn't be any directories in the "Image Directory".
-      .filter((file) => !usedUris.some((uri) => file.uri === uri)),
-    callback: (image) => deleteImage(image.uri),
-    onBatchComplete: (isFulfilled) => {
-      deletedCount += isFulfilled.length;
-    },
-  });
-
-  console.log(`Deleted ${deletedCount} unlinked images.`);
+//#region Internal Utils
+/** Iterate over a list of tracks, finding and saving its artwork. */
+async function saveSinglesArtwork(
+  singles: PartialTrack[],
+  onSave: (info: { artworkUri: string; trackId: string }) => Promise<void>,
+  options?: { endEarly?: boolean; onEndIteration?: VoidFunction },
+) {
+  for (const { id: trackId, uri } of singles) {
+    // Indicate we attempted to find artwork for a track. Do this before
+    // physically attempting to save the artwork in case an OOM error occurs,
+    // in which the app will essentially become "bricked".
+    await updateTrack(trackId, { fetchedArt: true });
+    const { uri: artworkUri } = await getArtworkUri(uri);
+    if (artworkUri) {
+      await onSave({ artworkUri, trackId });
+      onboardingStore.setState((prev) => ({ found: prev.found + 1 }));
+      if (options?.endEarly) return;
+    }
+    if (options?.onEndIteration) options.onEndIteration();
+  }
 }
 //#endregion
