+++ mobile/src/modules/scanning/helpers/rescan.ts
@@ -11,8 +11,9 @@ import { Resynchronize } from "~/stores/Playback/actions";
 import { clearAllQueries } from "~/lib/react-query";
 import { ToastOptions } from "~/lib/toast";
 import { wait } from "~/utils/promise";
-import { findAndSaveArtwork, cleanupImages } from "./artwork";
+import { findAndSaveArtwork } from "./artwork";
 import { cleanupDatabase, findAndSaveAudio } from "./audio";
+import { cleanupImages } from "./cleanup";
 import { savePathComponents } from "./folder";
 
 /** Look through our library for any new or updated tracks. */
